Limit number of generated suites per file
